CHANGE: Added a few additional things to make that faster
Use the streamID as JobGroup, and ONLY look at jobs belonging to a given
JobGroup (there is no need to monitor jobs for which the stream is
completed/closed/whatever the term is). The Java code should submit the
jobs using the streamID, and the monitoring loop should also use that
streamID. Makes life a lot faster.